+++ Studio/src/Application/Optimize/QOptimize.h
@@ -11,8 +11,8 @@ public:
   QOptimize(QObject* parent = nullptr,
     std::vector<ImageType::Pointer> inputs =
     std::vector<ImageType::Pointer>(), 
-    std::vector<std::array<itk::Point<float>, 3 > > cutPlanes =
-    std::vector<std::array<itk::Point<float>, 3 > >(),
+    std::vector<std::array<itk::Point<double>, 3 > > cutPlanes =
+    std::vector<std::array<itk::Point<double>, 3 > >(),
     size_t numScales = 1,
     std::vector<double> start_reg = std::vector<double>(),
     std::vector<double> end_reg = std::vector<double>(),
